+++ src/DependencyInjection/IconLoaderBundleExtension.php
@@ -2,7 +2,7 @@
 
 namespace Becklyn\IconLoader\DependencyInjection;
 
-use Becklyn\IconLoader\Loader\IconLoader;
+use Becklyn\IconLoader\Registry\IconRegistry;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
@@ -24,9 +24,16 @@ class IconLoaderBundleExtension extends Extension
 
         // map config to services
         $config = $this->processConfiguration(new IconLoaderBundleConfiguration(), $configs);
-        $directoriesGlob = $container->getParameter("kernel.project_dir") . "/" . \ltrim($config["search_glob"], "/");
-        $container
-            ->getDefinition(IconLoader::class)
-            ->setArgument('$directoriesGlob', $directoriesGlob);
+
+        $registry = $container->getDefinition(IconRegistry::class);
+
+        foreach ($config["namespaces"] as $namespace => $namespaceConfig)
+        {
+            $registry->addMethodCall("registerProjectNamespace", [
+                $namespace,
+                $namespaceConfig["path"],
+                $namespaceConfig["class_name"],
+            ]);
+        }
     }
 }