+++ src/Exception/DuplicateNamespaceException.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\IconLoader\Exception;
+
+class DuplicateNamespaceException extends IconLoaderException
+{
+}