+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-Changelog for 1.0.1
-===================
+2.0.0
+=====
+
+*   All icon paths are now namespaces.
+*   Added `RegisterIconNamespacesCompilerPass` to provide icons in bundles.
+*   All icons are now HTML wrapped.   
+
+
+1.0.1
+=====
 
 *   Bugfix: fixed the check whether a missing icon should throw.
+
+
+1.0.0
+=====
+
+Initial release `\o/`